hoka/src/plugins/bili_analysis/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-analysis-bilibili"
version = "2.1.6"
description = "nonebot2解析bilibili插件"
authors = ["mengshouer"]
license = "MIT"
readme = "README.md"

repository = "https://github.com/mengshouer/nonebot_plugin_analysis_bilibili"
keywords = ["nonebot", "nonebot2", "bilibili"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
lxml = "^4.6.2"
nonebot2 = "^2.0.0a8.post2"
nb-cli = "^0.3.1"
DateTime = "^4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.core.masonry.api"